refactor the database
