+++ network/src/main/java/com/nicos/network/data/repository_impl/PokemonListRepositoryImpl.kt
@@ -1,6 +1,8 @@
 package com.nicos.network.data.repository_impl
 
+import androidx.core.text.isDigitsOnly
 import com.nicos.database.data.room_database.entities.PokemonEntity
+import com.nicos.database.data.room_database.entities.toPokemonEntity
 import com.nicos.database.data.room_database.init_database.MyRoomDatabase
 import com.nicos.network.domain.remote.PokemonService
 import com.nicos.network.domain.repositories.PokemonListRepository
@@ -19,6 +21,12 @@ class PokemonListRepositoryImpl @Inject constructor(
     private val handlingError: HandlingError,
 ) : PokemonListRepository {
 
+    companion object {
+        private const val BASE_IMAGE_URL =
+            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/home/"
+        private const val PNG_FORMAT = ".png"
+    }
+
     override suspend fun fetchPokemonList(url: String?): Flow<Resource<MutableList<PokemonEntity>>> {
         return flow {
             try {
@@ -26,7 +34,8 @@ class PokemonListRepositoryImpl @Inject constructor(
                     if (url == null) pokemonService.getPokemon() else pokemonService.getPokemon(url)
                 val nextUrl = pokemonService.nextUrl
                 val pokemonEntityList = pokemonService.results
-                savePokemon(pokemonEntityList = pokemonEntityList)
+                savePokemon(pokemonEntityList = pokemonEntityList.map { it.toPokemonEntity() }
+                    .toMutableList())
 
                 emit(
                     Resource.Success(
@@ -40,11 +49,23 @@ class PokemonListRepositoryImpl @Inject constructor(
         }.flowOn(Dispatchers.IO)
     }
 
-    override suspend fun savePokemon(pokemonEntityList: MutableList<PokemonEntity>) =
-        PokemonEntity.savePokemonList(
-            pokemonEntityList = pokemonEntityList,
-            myRoomDatabase = myRoomDatabase
-        ).collect()
+    override suspend fun savePokemon(pokemonEntityList: MutableList<PokemonEntity>)  {
+        pokemonEntityList.forEach {
+            buildPokemonImageUrl(it)
+            if (it.imageUrl != null) {
+                myRoomDatabase.pokemonDao().insertOrReplaceObject(it)
+            }
+        }
+    }
+
+    private fun buildPokemonImageUrl(pokemonEntity: PokemonEntity) {
+        val pokemonIdAsString =
+            pokemonEntity.url?.substringBeforeLast("/")?.substringAfterLast("/")
+        if (pokemonIdAsString != null && pokemonIdAsString.isDigitsOnly()) {
+            pokemonEntity.order = pokemonIdAsString.toInt()
+            pokemonEntity.imageUrl = "${BASE_IMAGE_URL}$pokemonIdAsString${PNG_FORMAT}"
+        }
+    }
 
     override suspend fun offline(): Flow<Resource<MutableList<PokemonEntity>>> {
         return flow {
