+++ network/src/main/java/com/nicos/network/domain/remote/PokemonService.kt
@@ -1,7 +1,7 @@
 package com.nicos.network.domain.remote
 
-import com.nicos.database.data.room_database.entities.PokemonDetailsEntity
 import com.nicos.network.data.pokemon_response_model.PokemonResponse
+import com.nicos.network.domain.dto.PokemonDetailsDto
 import retrofit2.http.GET
 import retrofit2.http.Url
 
@@ -14,5 +14,5 @@ interface PokemonService {
     suspend fun getPokemon(@Url url: String): PokemonResponse
 
     @GET
-    suspend fun getPokemonDetails(@Url url: String): PokemonDetailsEntity
+    suspend fun getPokemonDetails(@Url url: String): PokemonDetailsDto
 }
